requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
streamlit>=1.22.0
openai>=1.2.0
python-dotenv>=1.0.0
pandas>=1.5.3
altair>=4.2.2
requests>=2.28.2
pydantic>=1.10.7
dataclasses-json>=0.5.7
motor>=4.1.0       # Async MongoDB driver
pymongo>=4.4.0     # MongoDB utilities (ObjectId, etc.)